[test] null_또는_빈문자()
