+++ src/main/java/utils/StringUtils.java
@@ -1,0 +1,8 @@
+package utils;
+
+public class StringUtils {
+
+    public static boolean isNullOrBlank(String value) {
+        return value == null || value.trim().isEmpty();
+    }
+}
